requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
coverage==7.0.0
Django==4.1.4
django-nose==1.4.7
djangorestframework==3.14.0
install==1.3.5
nose==1.3.7
psycopg==3.1.7
psycopg2==2.9.5
pytz==2022.7
sqlparse==0.4.3
typing-extensions==4.4.0
